[Feature43] DetailFragment에 있는 처리동작 viewmodel로 이동 1. image 공유 2. object detection
